being abstracting out TUI as a watcher
